total_score and total_weighting columns to FLOAT
--HG--
branch : 2.10

+++ modules/create_course/create_course_db.php
@@ -116,8 +116,8 @@ db_query("CREATE TABLE exercices (
       uid mediumint(8) NOT NULL default '0',
       RecordStartDate datetime NOT NULL default '0000-00-00',
       RecordEndDate datetime default NULL,
-      TotalScore int(11) NOT NULL default '0',
-      TotalWeighting int(11) default '0',
+      TotalScore float(5,2) NOT NULL default '0',
+      TotalWeighting float(5,2) default '0',
       attempt int(11) NOT NULL default '0',
       PRIMARY KEY  (eurid))
        $charset_spec");
