+++ modules/create_course/functions.php
@@ -184,8 +184,8 @@ function create_course_db($code)
                 uid mediumint(8) NOT NULL default '0',
                 RecordStartDate datetime NOT NULL default '0000-00-00',
                 RecordEndDate datetime default NULL,
-                TotalScore int(11) NOT NULL default '0',
-                TotalWeighting int(11) default '0',
+                TotalScore float(5,2) NOT NULL default '0',
+                TotalWeighting float(5,2) default '0',
                 attempt int(11) NOT NULL default '0',
                 PRIMARY KEY  (eurid))
                 $charset_spec");
